Remove NativeObjectWrapperBase, move logic to IUnknownComWrapper, fix finalization
